+++ frontend/frontend.config-cere-local.js
@@ -4,7 +4,7 @@ export const network = {
   tokenSymbol: 'CERE',
   tokenDecimals: 10,
   ss58Format: -1,
-  coinGeckoDenom: 'cere',
+  coinGeckoDenom: 'cere-network',
   nodeWs: 'wss://archive.mainnet.cere.network:9945',
   backendWs: 'ws://localhost:8082/v1/graphql',
   backendHttp: 'http://localhost:8080/api/v1',