+++ CHANGELOG.md
@@ -1,27 +1,35 @@
 # Cere Stats
 
 ## vNext
+
+- Show CERE-USD conversion rate in header
 - Updated crawlers log levels.
 
 ## v0.5.0
+
 - Added total transactions to homepage.
 - Updated tooltip position for blocks.
 
 ## v0.4.0
+
 - Updated logo and favicon.
 - Added functionality to filter blocks based on block hash.
 
 ## v0.3.1
+
 - Updated Cere Network Homepage link.
 
 ## v0.3.0
+
 - Added faucet functionality.
 - Added Google Analytics for PROD.
 - Updated backend config values.
 
 ## v0.2.0
+
 - Added UI to show EDP banner on homepage.
 
 ## v0.1.0
+
 - Adapted for Cere Network.
 - Added API to add and get EDP info.